requirements.txt
numpy>=1.24
pydantic>=1.10
PyYAML>=6.0
psutil>=5.9
